+++ packages/amplify-ui-angular/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.54](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.53...@aws-amplify/ui-angular@1.0.54) (2022-07-07)
+
+**Note:** Version bump only for package @aws-amplify/ui-angular
+
+
+
+
+
+## [1.0.53](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.52...@aws-amplify/ui-angular@1.0.53) (2022-06-18)
+
+**Note:** Version bump only for package @aws-amplify/ui-angular
+
+
+
+
+
+## [1.0.52](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.51...@aws-amplify/ui-angular@1.0.52) (2022-06-15)
+
+**Note:** Version bump only for package @aws-amplify/ui-angular
+
+
+
+
+
 ## [1.0.51](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-angular@1.0.50...@aws-amplify/ui-angular@1.0.51) (2022-05-24)
 
 **Note:** Version bump only for package @aws-amplify/ui-angular